requirements-streamlit.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-google-genai>=0.0.5
pydantic>=2.5.2
supabase>=2.0.3
python-dotenv>=1.0.0
google-generativeai>=0.3.1
streamlit==1.35.0 